requirements.txt
SpeechRecognition
googletrans
gtts
pygame
google.generativeai
